backend/requirements.txt
fastapi>=0.115.0
 uvicorn[standard]>=0.30.0
 python-multipart>=0.0.9
 pydantic>=2.8.2
 python-dotenv>=1.0.1
 langchain>=0.3.0
 langchain-core>=0.3.0
 langchain-community>=0.3.0
 langchain-openai>=0.2.0
 langchain-pinecone>=0.1.4
 langchain-text-splitters>=0.3.0
 openai>=1.47.0
 pinecone>=6.0.0,<7.0.0
 pypdf>=4.3.1